courses roles

+++ online_school_frontend/src/courses/courses.js
@@ -1,4 +1,12 @@
-import {getAuthHeaders, loadTeacher, closeModal, deleteRecord, updateRowCount} from "../app_funcs.js"
+import {
+    getAuthHeaders,
+    loadTeacher,
+    closeModal,
+    deleteRecord,
+    updateRowCount,
+    configureSchCourByRole,
+    getUserRole
+} from "../app_funcs.js"
 
 const apiUrl = "http://localhost:8086/courses";
 
@@ -10,7 +18,8 @@ export {
     saveEditedCourse,
     showAddForm,
     openEditModal,
-    filterAndSortCourses
+    filterAndSortCourses,
+    deleteCourse
 };
 
 window.fetchCourses = fetchCourses;
@@ -21,10 +30,15 @@ window.saveEditedCourse = saveEditedCourse;
 window.showAddForm = showAddForm;
 window.openEditModal = openEditModal;
 window.filterAndSortCourses = filterAndSortCourses;
+window.deleteCourse = deleteCourse;
 
 let coursesData = [];
 
 function fetchCourses() {
+    //отладка временно
+    const userRole = getUserRole();
+    console.log(userRole);
+
     fetch(apiUrl + "/", { headers: getAuthHeaders() })
         .then(response => {
             if (!response.ok) {
@@ -35,6 +49,7 @@ function fetchCourses() {
         .then(data => {
             coursesData = data;
             renderTable(coursesData);
+            configureSchCourByRole();
         })
         .catch(error => {
             console.error("Ошибка загрузки курсов:", error);
@@ -61,8 +76,8 @@ function renderTable(data) {
         <td>${course.description}</td>
         <td>${course.teacher_id.full_name}</td>
         <td>
-          <a href="#" onclick="openEditModal(${course.id})">Редактировать</a>
-          <a href="#" onclick="deleteCourse(${course.id})">Удалить</a>
+          <a href="#" class="can-edit" style="display: none;" onclick="openEditModal(${course.id})">Редактировать</a>
+          <a href="#" class="can-delete" style="display: none;" onclick="deleteCourse(${course.id})">Удалить</a>
         </td>
       </tr>
     `;
@@ -107,23 +122,27 @@ function showAddForm() {
     document.getElementById("course-name").value = "";
     document.getElementById("description").value = "";
     loadTeacher("teacher-id");
+
     document.getElementById("modal_add").style.display = "block";
     fetchCourses();
 }
 
-document.addEventListener("DOMContentLoaded", () => {
-    fetchCourses();
-});
 
 function saveCourses() {
-    const name = document.getElementById("course-name").value;
+    const userRole = getUserRole();
+    if (userRole !== "admin") {
+        alert("У вас нет прав на добавление расписания.");
+        return;
+    }
+
+    const course_name = document.getElementById("course-name").value;
     const description = document.getElementById("description").value;
-    const teacherId = document.getElementById("student-id").value;
+    const teacherId = document.getElementById("teacher-id").value;
 
     const newCourse = {
-        name,
+        course_name,
         description,
-        teacher: { id: parseInt(teacherId) },
+        teacher_id: { id: parseInt(teacherId) },
     };
 
     fetch(`${apiUrl}/new`, {
@@ -135,6 +154,7 @@ function saveCourses() {
             if (!response.ok) {
                 throw new Error("Ошибка добавления курса");
             }
+            alert("Новый курс добавлен!");
             closeModal("modal_add");
             fetchCourses();
         })
@@ -145,18 +165,36 @@ function saveCourses() {
 }
 
 function openEditModal(id) {
-    const course = coursesData.find(cour => cour.id === id);
-    if (!course) return;
+    const url = `${apiUrl}/edit/${id}`;
 
-    document.getElementById("edit-course-id").value = course.id;
-    document.getElementById("edit-course-name").value = course.course_name;
-    document.getElementById("edit-description").value = course.description;
-    loadTeacher("edit-teacher-id", course.teacher_id.id);
+    const userRole = getUserRole();
+    if (userRole !== "admin") {
+        alert("У вас нет прав на редактирование.");
+    } else {
+        fetch(url, {
+            method: "GET",
+            headers: getAuthHeaders(),
+        })
+            .then(response => {
+                if (!response.ok) {
+                    throw new Error("Ошибка получения данных для редактирования");
+                }
+                return response.json();
+            })
+            .then(course => {
+                document.getElementById("edit-course-id").value = course.id;
+                document.getElementById("edit-course-name").value = course.course_name;
+                document.getElementById("edit-description").value = course.description;
+                loadTeacher("edit-teacher-id", course.teacher_id.id);
 
-    document.getElementById("edit-modal").style.display = "block";
+                document.getElementById("edit-modal").style.display = "block";
+            })
+            .catch(error => console.error("Ошибка редактирования расписания: ", error));
+    }
 }
 
 function saveEditedCourse() {
+    const userRole = getUserRole();
     const courseId = document.getElementById("edit-course-id").value;
 
     const updatedCourse = {
@@ -165,25 +203,40 @@ function saveEditedCourse() {
         teacher_id: { id: document.getElementById("edit-teacher-id").value },
     };
 
-    fetch(`${apiUrl}/edit/${courseId}`, {
-        method: "PUT",
-        headers: getAuthHeaders(),
-        body: JSON.stringify(updatedCourse),
-    })
-        .then((response) => {
-            if (!response.ok) {
-                throw new Error("Ошибка обновления курса");
-            }
-            alert("Курс успешно обновлен!");
-            closeModal("edit-modal");
-            fetchCourses();
+    if (userRole !== "admin") {
+        alert("У вас нет прав на редактирование.");
+    } else {
+        fetch(`${apiUrl}/edit/${courseId}`, {
+            method: "PUT",
+            headers: getAuthHeaders(),
+            body: JSON.stringify(updatedCourse),
         })
-        .catch((error) => {
-            console.error("Ошибка обновления курса:", error);
-            alert("Ошибка сохранения изменений.");
-        });
+            .then((response) => {
+                if (!response.ok) {
+                    throw new Error("Ошибка обновления курса");
+                }
+                alert("Курс успешно обновлен!");
+                closeModal("edit-modal");
+                fetchCourses();
+            })
+            .catch((error) => {
+                console.error("Ошибка обновления курса:", error);
+                alert("Ошибка сохранения изменений.");
+            });
+    }
 }
 
-function deleteCourse(courseId) {
-    deleteRecord(id, apiUrl, fetchCourses);
+function deleteCourse(id) {
+    const userRole = getUserRole();
+    if (userRole !== "admin") {
+        alert("У вас нет прав на удаление курсов.");
+    } else {
+        deleteRecord(id, apiUrl, fetchCourses);
+    }
 }
+
+document.addEventListener("DOMContentLoaded", () => {
+    configureSchCourByRole();
+    fetchCourses();
+});
+
